deleted pdf and md files
